+++ workspace/cli/theme.py
@@ -17,7 +17,39 @@ class Colors:
     red = 167
 
 
-def echo(message=None, file=None, nl=True, err=True, color=None, rewrite=False):
+def _color_code(number: int):
+    assert number < 256
+    return f"\x1b[38;5;{number}m"
+
+
+TAGS = {
+    "header": _color_code(Colors.turquoise),
+    "success": _color_code(Colors.green),
+    "accent": _color_code(Colors.purple),
+    "warning": _color_code(Colors.yellow),
+    "error": _color_code(Colors.red),
+    "bold": _BOLD,
+}
+TAG_RE = "(" + "|".join([tag for tag in TAGS] + [tag[0] for tag in TAGS]) + ")"
+
+
+def echo(message: str, *args, file=None, nl=True, err=True, color=None, rewrite=False):
+    """Wrapper around `click.echo` which allows tag-based styling.
+
+    Supports %-style string formatting to escape text from external sources (whose 'tags'
+    should not be parsed).
+
+    Available tags are:
+    - <header> or <h>
+    - <success> or <s>
+    - <accent> or <a>
+    - <warning> or <w>
+    - <error> or <e>
+    - <bold> or <b>
+    """
+    if args:
+        args = tuple([escape(arg) for arg in args])
+        message = message % args
     message = colorize(message)
     if rewrite:
         message = _CLEAR + message
@@ -25,43 +57,59 @@ def echo(message=None, file=None, nl=True, err=True, color=None, rewrite=False):
 
 
 def colorize(message: str) -> str:
-    return "".join([token for token in _tokenize(message)])
+    return unescape("".join([token for token in _tokenize(message)]))
 
 
 def strip_tags(message: str) -> str:
-    return "".join([token for token in _tokenize(message, strip_tags=True)])
+    return unescape("".join([token for token in _tokenize(message, strip_tags=True)]))
+
+
+def escape(message: str) -> str:
+    """Escape tags which might be interpreted by the theme tokenizer.
+
+    Should be used when passing text from external sources to `theme.echo`.
+    """
+    return re.sub(
+        rf"<(/?{TAG_RE})>",
+        r"\<\1>",
+        message,
+    )
+
+
+def unescape(message: str) -> str:
+    return re.sub(
+        rf"\\<(/?{TAG_RE})>",
+        r"<\1>",
+        message,
+    )
 
 
 def _tokenize(string: str, strip_tags: bool = False):
-    style_tags = {
-        "header": Colors.turquoise,
-        "success": Colors.green,
-        "accent": Colors.purple,
-        "warning": Colors.yellow,
-        "error": Colors.red,
-    }
+    """Tokenizer for easy CLI theme tokenizer.
+
+    Parses HTML-inspired tags to describe the text style.
+    """
     tokens = {
-        "TAGOPEN": r"<[A-Za-z_-]+>",
-        "TAGCLOSE": r"</[(A-Za-z_-]+>",
+        "TAGOPEN": rf"(?<!\\)<{TAG_RE}>",
+        "TAGCLOSE": rf"(?<!\\)</{TAG_RE}>",
         "DEFAULT": r".",
     }
     token_regex = "|".join(f"(?P<{name}>{regex})" for name, regex in tokens.items())
     stack: List[str] = []
 
     def parse_tag(tag: str):
-        tags = set(style_tags) | {"bold"}
-        if tag in tags:
+        if tag in TAGS:
             return tag
-        for full_tag in tags:
+        for full_tag in TAGS:
             if full_tag.startswith(tag):
                 return full_tag
 
     def reset_code():
-        style = next((tag for tag in reversed(stack) if tag in style_tags), None)
+        style = next((tag for tag in reversed(stack) if tag != "bold"), None)
         bold = "bold" in stack
         result = _RESET
         if style:
-            result = result + _color_code(style_tags[style])
+            result = result + TAGS[style]
         if bold:
             result = result + _BOLD
         return result
@@ -73,18 +121,20 @@ def _tokenize(string: str, strip_tags: bool = False):
             if strip_tags:
                 continue
             tag = parse_tag(token[1:-1])
-            if tag == "bold":
-                yield _BOLD
-            elif tag in style_tags:
-                yield _color_code(style_tags[tag])
-            else:
-                assert False, f"Unknown token {token}"
+            try:
+                yield TAGS[tag]
+            except KeyError:
+                # Unknown tag - unlikely but would mean an error in TAG_RE
+                yield token
+                continue
             stack.append(tag)
         elif kind == "TAGCLOSE":
             if strip_tags:
                 continue
             tag = parse_tag(token[2:-1])
-            assert stack and (stack[-1] == tag), f"Invalid closing tag {token}"
+            if not stack or stack[-1] != tag:
+                yield token
+                continue
             stack.pop()
             yield reset_code()
         else:
@@ -94,14 +144,9 @@ def _tokenize(string: str, strip_tags: bool = False):
         yield reset_code()
 
 
-def _color_code(number: int):
-    assert number < 256
-    return f"\x1b[38;5;{number}m"
-
-
 def demo_theme():
     echo(
-        """<h><b>This is a header</b></h>
+        f"""<h><b>This is a header</b></h>
 
 This is some regular text, with <a>accented</a> parts.
 
@@ -111,6 +156,9 @@ Let's draw attention to the <s>good</s> things.
 
 <e>Something here has gone <b>very wrong</b>!</e>
 
+<Unknown> tags are ignored.
+
+{escape("Raw <b>text</b> can be escaped.")}
 """
     )  # pragma: no cover